Moves export functions into a seperate file
