+++ public/js/main.js
@@ -91,6 +91,57 @@
     }
   }
 
+  // src/export.js
+  function ExportRoute() {
+    let route = ``;
+    for (let i = 0; i < points.length - 1; i++) {
+      const angle1 = i === 0 ? 0 : CalculateAngle(points[i - 1], points[i]);
+      let angle2 = CalculateAngle(points[i], points[i + 1]);
+      if (points[i].d === -1) {
+        angle2 = FlipAngle(angle2);
+      }
+      let turn = CalculateTurn(angle1, angle2);
+      if (i > 0) {
+        if (points[i - 1].d === -1) {
+          turn *= -1;
+        }
+      }
+      const distance = CalculateDistance(points[i], points[i + 1]);
+      const func = points[i].f;
+      const direction = points[i].d;
+      route += `` + turn + `
+` + Math.round(angle2) + `
+` + direction + `
+` + Math.round(distance) + `
+` + func + `
+`;
+    }
+    return route.trim();
+  }
+  function DownloadRoute() {
+    let data = ExportRoute();
+    let blob = new Blob([data], { type: "text.plain" });
+    let url = URL.createObjectURL(blob);
+    menu.routeL.href = url;
+    menu.routeL.click();
+  }
+  function CopyRoute() {
+    navigator.clipboard.writeText(ExportRoute());
+  }
+  function ExportPoints() {
+    return JSON.stringify(points);
+  }
+  function DownloadPoints() {
+    let data = ExportPoints();
+    let blob = new Blob([data], { type: "application/json" });
+    let url = URL.createObjectURL(blob);
+    menu.pointsL.href = url;
+    menu.pointsL.click();
+  }
+  function CopyPoints() {
+    navigator.clipboard.writeText(ExportPoints());
+  }
+
   // src/main.js
   window.currentPoint = null;
   window.holdingPoint = null;
@@ -193,7 +244,7 @@
       MovePointDown();
     }
   };
-  var menu = {
+  window.menu = {
     formPoints: document.getElementById("point"),
     formSettings: document.getElementById("settings"),
     pointX: document.getElementById("PointX"),
@@ -310,55 +361,6 @@
   menu.pointsF.onchange = function() {
     ImportPointsFromFile();
   };
-  function ExportRoute() {
-    let route = ``;
-    for (let i = 0; i < points.length - 1; i++) {
-      const angle1 = i === 0 ? 0 : CalculateAngle(points[i - 1], points[i]);
-      let angle2 = CalculateAngle(points[i], points[i + 1]);
-      if (points[i].d === -1) {
-        angle2 = FlipAngle(angle2);
-      }
-      let turn = CalculateTurn(angle1, angle2);
-      if (i > 0) {
-        if (points[i - 1].d === -1) {
-          turn *= -1;
-        }
-      }
-      const distance = CalculateDistance(points[i], points[i + 1]);
-      const func = points[i].f;
-      const direction = points[i].d;
-      route += `` + turn + `
-` + Math.round(angle2) + `
-` + direction + `
-` + Math.round(distance) + `
-` + func + `
-`;
-    }
-    return route.trim();
-  }
-  function DownloadRoute() {
-    let data = ExportRoute();
-    let blob = new Blob([data], { type: "text.plain" });
-    let url = URL.createObjectURL(blob);
-    menu.routeL.href = url;
-    menu.routeL.click();
-  }
-  function CopyRoute() {
-    navigator.clipboard.writeText(ExportRoute());
-  }
-  function ExportPoints() {
-    return JSON.stringify(points);
-  }
-  function DownloadPoints() {
-    let data = ExportPoints();
-    let blob = new Blob([data], { type: "application/json" });
-    let url = URL.createObjectURL(blob);
-    menu.pointsL.href = url;
-    menu.pointsL.click();
-  }
-  function CopyPoints() {
-    navigator.clipboard.writeText(ExportPoints());
-  }
   function ImportPointsFromText() {
     try {
       json = JSON.parse(menu.pointsT.value);
